# openapi.yaml
openapi: 3.1.0
info:
  title: Pipeline AI DevOps Platform API
  version: 1.0.0
  description: |
    AI DevOps automation platform for deploying applications to Vercel and Render.
    Provides sandboxed workspace execution via Daytona, AI-powered repository analysis,
    and automated deployment orchestration.
  contact:
    name: Pipeline Support
    email: support@pipeline.dev
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://api.pipeline.dev/api/v1
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /deployments:
    get:
      summary: List all deployments
      tags: [Deployments]
      operationId: listDeployments
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, succeeded, failed, cancelled]
        - name: repo_id
          in: query
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of deployments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentList'

    post:
      summary: Create deployment plan
      description: Create a new deployment plan based on repository analysis
      tags: [Deployments]
      operationId: createDeployment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeploymentCreateRequest'
      responses:
        '201':
          description: Deployment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentPlan'
        '400':
          description: Invalid request
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DeploymentCreateRequest:
      type: object
      required: [repo_id, services]
      properties:
        repo_id:
          type: string
          format: uuid
        services:
          type: array
          items:
            $ref: '#/components/schemas/ServiceDeploymentConfig'
        environment:
          type: string
          enum: [development, staging, production]
          default: production
        branch:
          type: string
          description: Branch to deploy from
        env_variables:
          type: object
          additionalProperties:
            type: string
          description: Environment variables for all services
    ServiceDeploymentConfig:
      type: object
      required: [name, path, platform]
      properties:
        name:
          type: string
        path:
          type: string
        platform:
          type: string
          enum: [vercel, render, docker]
        build_command:
          type: string
        start_command:
          type: string
        output_directory:
          type: string
        env_variables:
          type: object
          additionalProperties:
            type: string
        vercel_config:
          $ref: '#/components/schemas/VercelDeploymentConfig'
        render_config:
          $ref: '#/components/schemas/RenderDeploymentConfig'
    VercelDeploymentConfig:
      type: object
      properties:
        project_name:
          type: string
          description: Vercel project name
        team_id:
          type: string
        framework:
          type: string
    RenderDeploymentConfig:
      type: object
      properties:
        service_name:
          type: string
        service_type:
          type: string
          enum: [web_service, static_site, background_worker, cron_job]
        plan:
          type: string
          default: starter
    DeploymentPlan:
      type: object
      required: [id, repo_id, status, services, created_at]
      properties:
        id:
          type: string
          format: uuid
        repo_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, planned, approved]
        services:
          type: array
          items:
            $ref: '#/components/schemas/PlannedService'
        environment:
          type: string
        branch:
          type: string
        created_at:
          type: string
          format: date-time
    PlannedService:
      type: object
      required: [name, path, platform]
      properties:
        name:
          type: string
        path:
          type: string
        platform:
          type: string
          enum: [vercel, render, docker]
        status:
          type: string
          enum: [pending, deploying, deployed, failed, skipped]
        estimated_duration_seconds:
          type: integer
          description: Estimated deployment duration
    DeploymentStatus:
      type: object
      required: [id, repo_id, status, services, created_at]
      properties:
        id:
          type: string
          format: uuid
        repo_id:
          type: string
          format: uuid
        sandbox_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, running, succeeded, failed, cancelled, retrying]
        services:
          type: array
          items:
            $ref: '#/components/schemas/ServiceDeployment'
        environment:
          type: string
        branch:
          type: string
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        duration_seconds:
          type: integer
        error_message:
          type: string
        retry_count:
          type: integer
          default: 0
    ServiceDeployment:
      type: object
      required: [name, path, platform, status]
      properties:
        name:
          type: string
        path:
          type: string
        platform:
          type: string
          enum: [vercel, render, docker]
        status:
          type: string
          enum: [pending, building, deploying, deployed, failed, skipped]
        build_logs_url:
          type: string
          format: uri
        deployment_url:
          type: string
          format: uri
        platform_deployment_id:
          type: string
          description: Deployment ID from the platform (Vercel/Render)
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        error_message:
          type: string
...
    DeploymentList:
      type: object
      required: [deployments, total, limit, offset]
      properties:
        deployments:
          type: array
          items:
            $ref: '#/components/schemas/DeploymentStatus'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
...